requirements.txt
# Core ML Dependencies
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.8

# Computer Vision
opencv-python>=4.9.0.80
Pillow>=10.1.0
ultralytics>=8.0.0

# Data Processing
numpy>=1.24.4
scikit-learn>=1.3.2
pandas>=2.0.0
scipy>=1.10.0

# Configuration & Utils
PyYAML>=6.0.1
tqdm>=4.66.1
requests>=2.31.0

# Text-to-Speech (for talking camera)
pyttsx3>=2.90

# Additional ML Utilities
huggingface-hub>=0.16.0
safetensors>=0.3.0

# Plotting & Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
